% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_in.R
\name{mm_to_in}
\alias{mm_to_in}
\title{Convert Millimeters to Inches}
\usage{
mm_to_in(mm)
}
\arguments{
\item{mm}{A numeric vector, with measurements in millimeters. Non-numeric values will result in an error.}
}
\value{
A numeric vector, with values converted from mm to inches. NA values remain NA.
}
\description{
This function converts measurements from millimeters (mm) to inches: 1 inch = 25.4 mm.
It is designed to ignore non-numeric inputs.
It will accept a numeric vector as input and return a vector of converted values.
}
\examples{
mm_to_in(25.4) # Should return 1
mm_to_in(c(25.4, 127)) # Should return 1 and 5
mm_to_in(NA) # Should return NA
}
